moved aux_funcs to utils
