+++ dnora/spectra1d/process.py
@@ -2,7 +2,7 @@ import numpy as np
 from abc import ABC, abstractmethod
 
 from dnora.type_manager.spectral_conventions import SpectralConvention
-from dnora.aux_funcs import check_that_spectra_are_consistent
+from dnora.utils.spec import check_that_spectra_are_consistent
 
 
 class SpectralProcessor(ABC):
